+++ thermal/__init__.py
@@ -4,5 +4,5 @@ __author__ = """Thijs van den Berg"""
 __email__ = 'thijs@sitmo.com'
 __version__ = '0.6.2'
 
-from thermal.resample import Resample, ResampleGmm, ResampleHist, ResampleKde
+from thermal.sample import Sample, SampleGmm, SampleHist, SampleKde
 from thermal.stats import anderson_test, ks_test