+++ thermal/sample/__init__.py
@@ -1,0 +1,6 @@
+"""The sample module implements sampling algorithms."""
+
+from thermal.sample.generic import Sample
+from thermal.sample.gmm import SampleGmm
+from thermal.sample.hist import SampleHist
+from thermal.sample.kde import SampleKde